Move error code to error module
In preparation for adding a bunch more error types move the current ones
out of `parse` into `error`.
